requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
langchain-community
sentence-transformers
pydantic
python-multipart
docx2txt
PyMuPDF
supabase
